requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyyaml>=6.0
asyncio-mqtt>=0.13.0
websockets>=11.0
aiohttp>=3.8.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Data analysis and ML
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Database and storage
sqlite3
sqlalchemy>=2.0.0
alembic>=1.11.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Logging and observability
langfuse>=2.0.0
structlog>=23.1.0

# Development tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional: LLM integration
openai>=1.0.0
langchain>=0.1.0
